requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.7
djangorestframework==3.14.0
filelock==3.13.1
pbr==6.0.0
platformdirs==4.0.0
pytz==2023.3.post1
sqlparse==0.4.4
stevedore==5.1.0
tzdata==2023.3
virtualenv==20.24.7
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
